AshAuthentication.AddOn.Confirmation.Actions (ash_authentication v4.0.4)

Actions for the confirmation add-on.

Provides the code interface for working with resources via confirmation.

Functions

confirm(strategy, params, opts \\ [])

@spec confirm(AshAuthentication.AddOn.Confirmation.t(), map(), keyword()) ::
  {:ok, Ash.Resource.record()} | {:error, any()}

Attempt to confirm a user.

get_changes(strategy, jti, opts \\ [])

@spec get_changes(AshAuthentication.AddOn.Confirmation.t(), String.t(), keyword()) ::
  {:ok, map()} | :error

Get changes from the tokens resource for application.

store_changes(strategy, token, changeset, opts \\ [])

@spec store_changes(
  AshAuthentication.AddOn.Confirmation.t(),
  String.t(),
  Ash.Changeset.t(),
  keyword()
) ::
  :ok | {:error, any()}

Store changes in the tokens resource for later re-use.