AshAuthentication.TokenResource.Info (ash_authentication v4.0.4)

Introspection functions for the AshAuthentication.TokenResource Ash extension.

Functions

token_confirmation_get_changes_action_name(dsl_or_extended)

@spec token_confirmation_get_changes_action_name(dsl_or_extended :: module() | map()) ::
  {:ok, atom()} | :error

The name of the action used to get confirmation changes.

token_confirmation_get_changes_action_name!(dsl_or_extended)

@spec token_confirmation_get_changes_action_name!(dsl_or_extended :: module() | map()) ::
  atom() | no_return()

The name of the action used to get confirmation changes.

token_confirmation_options(dsl_or_extended)

@spec token_confirmation_options(dsl_or_extended :: module() | map()) :: %{
  required(atom()) => any()
}

token.confirmation DSL options

Returns a map containing the and any configured or default values.

token_confirmation_store_changes_action_name(dsl_or_extended)

@spec token_confirmation_store_changes_action_name(
  dsl_or_extended :: module() | map()
) ::
  {:ok, atom()} | :error

The name of the action used to store confirmation changes.

token_confirmation_store_changes_action_name!(dsl_or_extended)

@spec token_confirmation_store_changes_action_name!(
  dsl_or_extended :: module() | map()
) ::
  atom() | no_return()

The name of the action used to store confirmation changes.

token_domain(dsl_or_extended)

@spec token_domain(dsl_or_extended :: module() | map()) :: {:ok, module()} | :error

The Ash domain to use to access this resource.

token_domain!(dsl_or_extended)

@spec token_domain!(dsl_or_extended :: module() | map()) :: module() | no_return()

The Ash domain to use to access this resource.

token_expunge_expired_action_name(dsl_or_extended)

@spec token_expunge_expired_action_name(dsl_or_extended :: module() | map()) ::
  {:ok, atom()} | :error

The name of the action used to remove expired tokens.

token_expunge_expired_action_name!(dsl_or_extended)

@spec token_expunge_expired_action_name!(dsl_or_extended :: module() | map()) ::
  atom() | no_return()

The name of the action used to remove expired tokens.

token_expunge_interval(dsl_or_extended)

@spec token_expunge_interval(dsl_or_extended :: module() | map()) ::
  {:ok, pos_integer()} | :error

How often to scan this resource for records which have expired, and thus can be removed.

token_expunge_interval!(dsl_or_extended)

@spec token_expunge_interval!(dsl_or_extended :: module() | map()) ::
  pos_integer() | no_return()

How often to scan this resource for records which have expired, and thus can be removed.

token_get_token_action_name(dsl_or_extended)

@spec token_get_token_action_name(dsl_or_extended :: module() | map()) ::
  {:ok, atom()} | :error

The name of the action used to retrieve tokens from the store, if require_tokens_for_authentication? is enabled in your authentication resource.

token_get_token_action_name!(dsl_or_extended)

@spec token_get_token_action_name!(dsl_or_extended :: module() | map()) ::
  atom() | no_return()

The name of the action used to retrieve tokens from the store, if require_tokens_for_authentication? is enabled in your authentication resource.

token_options(dsl_or_extended)

@spec token_options(dsl_or_extended :: module() | map()) :: %{
  required(atom()) => any()
}

token DSL options

Returns a map containing the and any configured or default values.

token_read_expired_action_name(dsl_or_extended)

@spec token_read_expired_action_name(dsl_or_extended :: module() | map()) ::
  {:ok, atom()} | :error

The name of the action use to find all expired tokens.

token_read_expired_action_name!(dsl_or_extended)

@spec token_read_expired_action_name!(dsl_or_extended :: module() | map()) ::
  atom() | no_return()

The name of the action use to find all expired tokens.

token_revocation_is_revoked_action_name(dsl_or_extended)

@spec token_revocation_is_revoked_action_name(dsl_or_extended :: module() | map()) ::
  {:ok, atom()} | :error

The name of the action used to check if a token is revoked.

token_revocation_is_revoked_action_name!(dsl_or_extended)

@spec token_revocation_is_revoked_action_name!(dsl_or_extended :: module() | map()) ::
  atom() | no_return()

The name of the action used to check if a token is revoked.

token_revocation_options(dsl_or_extended)

@spec token_revocation_options(dsl_or_extended :: module() | map()) :: %{
  required(atom()) => any()
}

token.revocation DSL options

Returns a map containing the and any configured or default values.

token_revocation_revoke_token_action_name(dsl_or_extended)

@spec token_revocation_revoke_token_action_name(dsl_or_extended :: module() | map()) ::
  {:ok, atom()} | :error

The name of the action used to revoke tokens.

token_revocation_revoke_token_action_name!(dsl_or_extended)

@spec token_revocation_revoke_token_action_name!(dsl_or_extended :: module() | map()) ::
  atom() | no_return()

The name of the action used to revoke tokens.

token_store_token_action_name(dsl_or_extended)

@spec token_store_token_action_name(dsl_or_extended :: module() | map()) ::
  {:ok, atom()} | :error

The name of the action to use to store a token, if require_tokens_for_authentication? is enabled in your authentication resource.

token_store_token_action_name!(dsl_or_extended)

@spec token_store_token_action_name!(dsl_or_extended :: module() | map()) ::
  atom() | no_return()

The name of the action to use to store a token, if require_tokens_for_authentication? is enabled in your authentication resource.